requirements.txt
streamlit
requests
pdfminer.six
python-docx
